refactor: Update BankRepository
Add interface BankRepository.
