+++ src/main/java/com/example/grpcpractice/server/repository/BankRepository.java
@@ -1,0 +1,9 @@
+package com.example.grpcpractice.server.repository;
+
+public interface BankRepository {
+    int findBalanceById(int accountId);
+
+    Integer deposit(int accountId, int amount);
+
+    Integer withdraw(int accountId, int amount);
+}
